requirements.txt
pandas
azure-eventhub==5.11.4